+++ src/datasets/domain/models/ExportedDatasetMetadata.ts
@@ -2,8 +2,3 @@ export type ExportedDatasetMetadata = {
   content: string
   contentType: string
 }
-
-export enum DatasetMetadataExportVersion {
-  LATEST_PUBLISHED = ':latest-published',
-  DRAFT = ':draft'
-}