+++ src/datasets/domain/repositories/IDatasetsRepository.ts
@@ -20,7 +20,7 @@ import { StorageDriver } from '../../../core/domain/models/StorageDriver'
 import { DatasetUploadLimits } from '../models/DatasetUploadLimits'
 import { DatasetReview } from '../models/DatasetReview'
 import { ExportedDatasetMetadata } from '../models/ExportedDatasetMetadata'
-import { DatasetMetadataExportVersion } from '../models/ExportedDatasetMetadata'
+import { DatasetNotNumberedVersion } from '../models/DatasetNotNumberedVersion'
 
 export interface IDatasetsRepository {
   getDataset(
@@ -91,7 +91,7 @@ export interface IDatasetsRepository {
   exportDatasetMetadata(
     datasetId: number | string,
     exporter: string,
-    version?: DatasetMetadataExportVersion
+    version?: DatasetNotNumberedVersion.LATEST_PUBLISHED | DatasetNotNumberedVersion.DRAFT
   ): Promise<ExportedDatasetMetadata>
   getDatasetAvailableDatasetTypes(): Promise<DatasetType[]>
   getDatasetAvailableDatasetType(datasetTypeId: number | string): Promise<DatasetType>